+++ Source code/sfpca/.Rproj.user/A8A22B62/pcs/source-pane.pper
@@ -1,3 +1,3 @@
 {
-    "activeTab" : 0
+    "activeTab" : -1
 }
