+++ Source code/sfpca/.Rproj.user/A8A22B62/pcs/windowlayoutstate.pper
@@ -1,8 +1,8 @@
 {
     "left" : {
         "panelheight" : 628,
-        "splitterpos" : 376,
-        "topwindowstate" : "NORMAL",
+        "splitterpos" : 263,
+        "topwindowstate" : "HIDE",
         "windowheight" : 666
     },
     "right" : {
