+++ Source code/sfpca/.Rproj.user/A8A22B62/sdb/prop/INDEX
@@ -1,4 +1,6 @@
 ~%2FDesktop%2FMoMA-prototype%2FSource%20code%2Fsfpca%2FNAMESPACE="F9B1123E"
 ~%2FDesktop%2FMoMA-prototype%2FSource%20code%2Fsfpca%2FR%2FSFPCA%20demo.R="4897A42B"
 ~%2FDesktop%2FMoMA-prototype%2FSource%20code%2Fsfpca%2Fsrc%2Fcommon.h="3DBF5828"
+~%2FDesktop%2FMoMA-prototype%2FSource%20code%2Fsfpca%2Fsrc%2Fmodel.cpp="3649040"
+~%2FDesktop%2FMoMA-prototype%2FSource%20code%2Fsfpca%2Fsrc%2Fmodel.h="56C6958"
 ~%2FDesktop%2FMoMA-prototype%2FSource%20code%2Fsfpca%2Fsrc%2Fsfpca_extended.cpp="C0292BC2"
