+++ Source code/sfpca/.Rproj.user/shared/notebooks/paths
@@ -1,2 +1,5 @@
 C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/NAMESPACE="8B2182B3"
+C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/src/common.h="69CB1A93"
+C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/src/model.cpp="90D0C547"
+C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/src/model.h="D90A198C"
 C:/Users/Smart/Desktop/MoMA-prototype/Source code/sfpca/src/sfpca_extended.cpp="C69A9E1F"
