+++ Source code/sfpca/src/common.h
@@ -11,5 +11,18 @@ arma::vec VCONST_VEC;
 using namespace std;
 using namespace Rcpp;
 using namespace arma;
+const bool DEBUG = 0;
 
+int myassert(bool flag, std::string info)
+{
+    if (flag == 0)
+    {
+        throw std::invalid_argument(info);
+    }
+    return 0;
+};
+double mat_norm(arma::vec u, arma::mat S_u)
+{
+    return sqrt(as_scalar(u.t() * S_u * u));
+}
 #endif
